+++ lib/crucible/safetensors/vector_inspect.ex
@@ -1,0 +1,5 @@
+defmodule Crucible.Safetensors.VectorInspect do
+  @moduledoc "Compatibility namespace for `CrucibleSafetensors.VectorInspect`."
+
+  defdelegate inspect(path, opts \\ []), to: CrucibleSafetensors.VectorInspect
+end